Move KeyStrokeLog out of normal_mode
